+++ HISTORY.md
@@ -4,6 +4,8 @@
 
 ### Changelog
 
+- Added chunk computation in CVT brute force calculation
+
 #### API
 
 - Add GradientOperatorEmitter to support OMG-MEGA and OG-MAP-Elites ({pr}`348`)